Feature 018: Race Outcome Stat Progression System (#20)
* Phase 1: Career Phase System with Stat Progression Configuration

Implemented Feature 018 Phase 1 - Career phase multipliers for stat progression:
- Young horses (0-9 races): 0.80x efficiency
- Prime horses (10-29 races): 1.20x efficiency
- Veteran horses (30-49 races): 0.60x efficiency
- Old horses (50+ races): 0.20x efficiency

Added comprehensive configuration constants for all 6 phases:
- Base stat growth rates and career phase boundaries
- Race-type focus multipliers (Sprint/Distance/Classic)
- Performance bonuses (Win/Place/Show/Mid-pack/Back of pack)
- Happiness change values

Created StatProgressionCalculator with CalculateAgeMultiplier method
and full test coverage (boundary values and career phase transitions).

Reorganized Racing service structure:
- Moved calculators to Calculators/ folder
- Moved configuration to Config/ folder
- Moved core racing classes to service root
- Updated all test references to new structure

* Phase 2: Core Stat Growth Formula with Career Phase Integration

Implemented GrowStat method with base 2% growth formula:
- Gap-based growth: (DominantPotential - Actual) × 0.02 × CareerMultiplier
- Ceiling enforcement: no growth when Actual >= DominantPotential
- Career phase integration: uses multipliers from Phase 1
- Automatic slowdown as horses approach genetic ceiling

Added comprehensive test coverage:
- Basic growth calculations for all career phases
- Ceiling enforcement (zero growth at cap)
- Near-ceiling behavior (diminishing returns)
- Large gap proportional growth
- Theory tests for boundary conditions

Fixed namespace references across test suite:
- Updated all test files to use new Calculators namespace
- Fixed RaceExecutor.cs unused namespace import
- All tests passing (13 tests in StatProgressionCalculatorTests)

Phase 2 delivers end-to-end stat growth capability with
full TDD coverage and integration with Phase 1 career system.

* Phase 3: Performance Bonus System

Implemented CalculatePerformanceMultiplier to reward competitive performance:
- Win (1st): 1.50x multiplier - winners learn fastest
- Place (2nd): 1.25x multiplier - strong performance bonus
- Show (3rd): 1.10x multiplier - modest performance boost
- Mid-pack (4th-6th): 1.00x multiplier - baseline learning
- Back of pack (7th+): 0.75x penalty - poor performance reduces growth

Added comprehensive test coverage:
- Individual tests for each position category
- Theory tests for boundary conditions
- Small field scenarios (3-horse races)
- Large field scenarios (10-horse races)
- All 24 tests passing

Phase 3 delivers competitive incentive system encouraging
horse owners to race in appropriate competition levels.

* Phase 4: Race-Type Stat Focus System

Implemented race distance-based stat specialization:
- Sprint races (≤6f): Speed 1.50x, Agility 1.25x, others 0.75x
  Develops speed demons for short bursts
- Distance races (≥11f): Stamina 1.50x, Durability 1.25x, others 0.75x
  Develops endurance horses for marathons
- Classic races (7-10f): All stats 1.00x
  Provides balanced, well-rounded development

Created RaceTypeFocusMultipliers record struct to encapsulate
stat-specific multipliers in a type-safe, immutable way.

Added comprehensive test coverage:
- Individual tests for Sprint, Distance, and Classic races
- Theory tests for distance boundaries (6f, 7f, 10f, 11f)
- All 31 tests passing (Phases 1-4 complete)

Phase 4 enables strategic race selection - owners can
specialize horses by choosing appropriate race types.

* Feature 018: Race Outcome Stat Progression - Phases 5 & 6

Integrated stat progression system into race execution flow and
implemented happiness system based on race performance.

Phase 5: RaceExecutor Integration
- Created IStatProgressionCalculator interface for better testability
- Injected IStatProgressionCalculator into RaceExecutor via DI
- Added ApplyStatProgression method to process all horses after race
- Implemented ApplyStatGrowth helper for individual stat updates
- Modified DetermineRaceResults from static to instance method
- Applied stat progression for Speed, Agility, Stamina, and Durability
- Enforced genetic ceiling: stats cannot exceed DominantPotential

Phase 6: Happiness System
- Implemented ApplyHappinessChange method
- Position-based happiness changes:
  * Win: +8, Place: +4, Show: +2
  * Mid-pack (4-6): 0, Back of pack (7+): -3
- Added exhaustion penalty: -5 happiness when stamina ≤ 10%
- Clamped happiness to 0-100 range

Updated 5 test files to include StatProgressionCalculator dependency
in RaceExecutor constructor calls. All tests passing (59 tests).

* Convert horse statistics from byte to double precision

Fixed critical truncation bug where stat growth was being lost due to
byte casting. Stats now properly preserve fractional values enabling
gradual progression over multiple races.

Changes:
- HorseStatistic: Actual, DominantPotential, RecessivePotential now double
- Horse entity: All stat properties (Speed/Stamina/Agility/Durability/Happiness) return double
- StatProgressionCalculator: GrowStat uses double parameters
- RaceRunHorse: InitialStamina changed to double for consistency
- BreedingExecutor: Updated stat generation to use double
- All calculators: Updated to accept double stat values
- Test suite: Updated 517 tests to use double precision

Impact:
- Stat growth now works correctly (0.80, 1.92 preserved vs truncated to 0, 1)
- Horses progress realistically toward genetic ceilings over career
- All 517 unit tests passing

+++ TripleDerby.Tests.Unit/Breeding/GenerateHorseStatisticsTests.cs
@@ -180,8 +180,8 @@ public class GenerateHorseStatisticsTests
 
             // Because our mock returns 'min' for actual selection,
             // expectedActual = Math.Max(1, dominantPotential / 3)
-            byte expectedDominant = hs.DominantPotential;
-            int expectedMin = Math.Max(1, expectedDominant / 3);
+            double expectedDominant = hs.DominantPotential;
+            double expectedMin = Math.Max(1, expectedDominant / 3);
             Assert.Equal((byte)expectedMin, hs.Actual);
         }
     }
